refactor(pretty_json): adds a field to existing json codec instead of a new pretty_json codec
