+++ changelog.d/20384_pretty_json_config.enhancement.md
@@ -1,3 +1,3 @@
-Add `pretty_json` encoding codec to output a prettified json format.
+Add `use_pretty_json` option to json coded to output a prettified json format.
 
 authors: lsampras
