+++ lib/k8s-e2e-tests/tests/vector-agent.rs
@@ -993,6 +993,8 @@ async fn custom_selectors() -> Result<(), Box<dyn std::error::Error>> {
               inputs: [kubernetes_logs]
               encoding:
                 codec: json
+                json:
+                    use_pretty_json: false
     "#};
 
     let vector = framework
@@ -1395,6 +1397,8 @@ async fn glob_pattern_filtering() -> Result<(), Box<dyn std::error::Error>> {
               inputs: [kubernetes_logs]
               encoding:
                 codec: json
+                json:
+                    use_pretty_json: false
     "#};
 
     let vector = framework
