+++ src/codecs/encoding/encoder.rs
@@ -106,15 +106,11 @@ impl Encoder<Framer> {
     /// Get the HTTP content type.
     pub const fn content_type(&self) -> &'static str {
         match (&self.serializer, &self.framer) {
+            (Serializer::Json(_) | Serializer::NativeJson(_), Framer::NewlineDelimited(_)) => {
+                "application/x-ndjson"
+            }
             (
-                Serializer::Json(_) | Serializer::PrettyJson(_) | Serializer::NativeJson(_),
-                Framer::NewlineDelimited(_),
-            ) => "application/x-ndjson",
-            (
-                Serializer::Gelf(_)
-                | Serializer::Json(_)
-                | Serializer::PrettyJson(_)
-                | Serializer::NativeJson(_),
+                Serializer::Gelf(_) | Serializer::Json(_) | Serializer::NativeJson(_),
                 Framer::CharacterDelimited(CharacterDelimitedEncoder { delimiter: b',' }),
             ) => "application/json",
             (Serializer::Native(_), _) | (Serializer::Protobuf(_), _) => "application/octet-stream",
@@ -125,7 +121,6 @@ impl Encoder<Framer> {
                 | Serializer::Json(_)
                 | Serializer::Logfmt(_)
                 | Serializer::NativeJson(_)
-                | Serializer::PrettyJson(_)
                 | Serializer::RawMessage(_)
                 | Serializer::Text(_),
                 _,
