+++ src/sinks/datadog/logs/config.rs
@@ -186,7 +186,7 @@ mod test {
     use crate::codecs::EncodingConfigWithFraming;
     use crate::components::validation::prelude::*;
     use vector_lib::{
-        codecs::{JsonSerializerConfig, MetricTagValues},
+        codecs::{encoding::format::JsonSerializerOptions, JsonSerializerConfig, MetricTagValues},
         config::LogNamespace,
     };
 
@@ -209,7 +209,8 @@ mod test {
 
             let encoding = EncodingConfigWithFraming::new(
                 None,
-                JsonSerializerConfig::new(MetricTagValues::Full).into(),
+                JsonSerializerConfig::new(MetricTagValues::Full, JsonSerializerOptions::default())
+                    .into(),
                 config.encoding.clone(),
             );
