+++ src/sinks/splunk_hec/logs/config.rs
@@ -291,7 +291,7 @@ mod tests {
     use super::*;
     use crate::components::validation::prelude::*;
     use vector_lib::{
-        codecs::{JsonSerializerConfig, MetricTagValues},
+        codecs::{encoding::format::JsonSerializerOptions, JsonSerializerConfig, MetricTagValues},
         config::LogNamespace,
     };
 
@@ -316,7 +316,11 @@ mod tests {
                 sourcetype: None,
                 source: None,
                 encoding: EncodingConfig::new(
-                    JsonSerializerConfig::new(MetricTagValues::Full).into(),
+                    JsonSerializerConfig::new(
+                        MetricTagValues::Full,
+                        JsonSerializerOptions::default(),
+                    )
+                    .into(),
                     Transformer::default(),
                 ),
                 compression: Compression::default(),
